requirements.txt
click==8.1.0
DateTime==4.4
Flask==2.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
pytz==2022.1
SQLAlchemy==1.4.32
Werkzeug==2.0.0
zope.interface==5.4.0

python-dotenv==0.20.0